requirements.txt
praw
openai
pydantic
python-dotenv
requests
beautifulsoup4
tqdm